refactor: split main into smaller modules
